+++ hack/.golint_failures
@@ -813,6 +813,7 @@ test/e2e/kubectl
 test/e2e/lifecycle
 test/e2e/lifecycle/bootstrap
 test/e2e/metrics
+test/e2e/network
 test/e2e/scalability
 test/e2e/scheduling
 test/e2e/storage
