+++ test/e2e/BUILD
@@ -28,6 +28,7 @@ go_test(
         "//test/e2e/lifecycle:go_default_library",
         "//test/e2e/lifecycle/bootstrap:go_default_library",
         "//test/e2e/metrics:go_default_library",
+        "//test/e2e/network:go_default_library",
         "//test/e2e/scalability:go_default_library",
         "//test/e2e/scheduling:go_default_library",
         "//test/e2e/storage:go_default_library",
@@ -52,26 +53,18 @@ go_library(
         "audit.go",
         "certificates.go",
         "dashboard.go",
-        "dns.go",
-        "dns_common.go",
-        "dns_configmap.go",
         "e2e.go",
         "events.go",
         "example_cluster_dns.go",
         "examples.go",
-        "firewall.go",
         "generated_clientset.go",
         "gke_local_ssd.go",
         "gke_node_pools.go",
         "ingress.go",
-        "kube_proxy.go",
         "kubelet.go",
         "kubelet_perf.go",
         "limit_range.go",
         "network_partition.go",
-        "network_policy.go",
-        "networking.go",
-        "networking_perf.go",
         "no-snat.go",
         "nodeoutofdisk.go",
         "pod_gc.go",
@@ -80,30 +73,22 @@ go_library(
         "pre_stop.go",
         "resource_quota.go",
         "security_context.go",
-        "service.go",
         "service_accounts.go",
         "service_latency.go",
         "serviceloadbalancers.go",
         "ssh.go",
         "ubernetes_lite.go",
-        "util_iperf.go",
     ],
     tags = ["automanaged"],
     deps = [
         "//pkg/api:go_default_library",
         "//pkg/api/testapi:go_default_library",
         "//pkg/api/v1/pod:go_default_library",
-        "//pkg/api/v1/service:go_default_library",
-        "//pkg/apis/networking:go_default_library",
-        "//pkg/client/clientset_generated/internalclientset:go_default_library",
-        "//pkg/cloudprovider:go_default_library",
         "//pkg/cloudprovider/providers/azure:go_default_library",
         "//pkg/cloudprovider/providers/gce:go_default_library",
-        "//pkg/controller/endpoint:go_default_library",
         "//pkg/controller/node:go_default_library",
         "//pkg/kubelet/apis:go_default_library",
         "//pkg/kubelet/apis/stats/v1alpha1:go_default_library",
-        "//pkg/master/ports:go_default_library",
         "//pkg/quota/evaluator/core:go_default_library",
         "//pkg/util/logs:go_default_library",
         "//pkg/util/version:go_default_library",
@@ -115,7 +100,6 @@ go_library(
         "//test/e2e/manifest:go_default_library",
         "//test/e2e/metrics:go_default_library",
         "//test/e2e_federation:go_default_library",
-        "//test/images/net/nat:go_default_library",
         "//test/utils:go_default_library",
         "//vendor/github.com/golang/glog:go_default_library",
         "//vendor/github.com/google/cadvisor/info/v1:go_default_library",
@@ -152,6 +136,8 @@ go_library(
     ],
 )
 
+# This is a handwritten rule. Do not delete, it will not be regenerated by
+# update-bazel.sh.
 genrule(
     name = "gen_e2e.test",
     testonly = 1,
@@ -184,6 +170,7 @@ filegroup(
         "//test/e2e/lifecycle:all-srcs",
         "//test/e2e/manifest:all-srcs",
         "//test/e2e/metrics:all-srcs",
+        "//test/e2e/network:all-srcs",
         "//test/e2e/perftype:all-srcs",
         "//test/e2e/scalability:all-srcs",
         "//test/e2e/scheduling:all-srcs",
