+++ test/e2e/network/OWNERS
@@ -1,0 +1,4 @@
+reviewers:
+  - sig-network-reviewers
+approvers:
+  - sig-network-approvers
