+++ test/e2e/network/util_iperf.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package e2e
+package network
 
 // Tests network performance using iperf or other containers.
 import (
